Fix signed - unsigned comparison warning

+++ DynamicHooks/conventions/x86MsCdecl.cpp
@@ -127,7 +127,7 @@ int x86MsCdecl::GetArgRegisterSize()
 	return iArgRegisterSize;
 }
 
-void* x86MsCdecl::GetArgumentPtr(int iIndex, CRegisters* pRegisters)
+void* x86MsCdecl::GetArgumentPtr(unsigned int iIndex, CRegisters* pRegisters)
 {
 	if (iIndex < 0 || iIndex >= m_vecArgTypes.length())
 		return NULL;
@@ -142,8 +142,8 @@ void* x86MsCdecl::GetArgumentPtr(int iIndex, CRegisters* pRegisters)
 		return pRegister->m_pAddress;
 	}
 
-	int iOffset = 4;
-	for(int i=0; i < iIndex; i++)
+	size_t iOffset = 4;
+	for(unsigned int i=0; i < iIndex; i++)
 	{
 		if (m_vecArgTypes[i].custom_register == None)
 			iOffset += m_vecArgTypes[i].size;
@@ -152,7 +152,7 @@ void* x86MsCdecl::GetArgumentPtr(int iIndex, CRegisters* pRegisters)
 	return (void *) (pRegisters->m_esp->GetValue<unsigned long>() + iOffset);
 }
 
-void x86MsCdecl::ArgumentPtrChanged(int iIndex, CRegisters* pRegisters, void* pArgumentPtr)
+void x86MsCdecl::ArgumentPtrChanged(unsigned int iIndex, CRegisters* pRegisters, void* pArgumentPtr)
 {
 }
